Move callbacks testing code into its own fixture crate for testing
