+++ examples/callbacks/uniffi.toml
@@ -1,0 +1,3 @@
+[bindings.kotlin]
+package_name = "uniffi.callbacks"
+cdylib_name = "callbacks"
