+++ fixtures/callbacks/tests/test_generated_bindings.rs
@@ -1,0 +1,4 @@
+uniffi_macros::build_foreign_language_testcases!(
+    "src/callbacks.udl",
+    ["tests/bindings/test_callbacks.kts"]
+);
